Retain fresh methods inside other JS fresh methods
Previously, in JavaScript, when a fresh method was located inside
another fresh method the last element of the body (the return value)
would be lost for the inheritance case, because it was removed in-place
to be compiled separately. Only nested fresh methods evinced this
behaviour as in any other case the method body would never be used
again. This commit restores the last element of the body as soon as the
body of the fresh method has been compiled.